feat: aux deployments and superpool tests

+++ script/oracle/DeployHyperliquidOracle.s.sol
@@ -7,9 +7,9 @@ import { HyperliquidOracle } from "src/oracle/HyperliquidOracle.sol";
 
 contract DeployHyperliquidOracle is BaseScript {
     address public immutable asset = 0x9cf99220F7dA086048D9fd4455407Ca8D65A6588;
-    uint public immutable assetIndex = 125;
-    uint public immutable assetAmtScale = 1;
-    uint public immutable assetPriceScale = 1e12;
+    uint256 public immutable assetIndex = 125;
+    uint256 public immutable assetAmtScale = 1;
+    uint256 public immutable assetPriceScale = 1e12;
 
     HyperliquidOracle oracle;
     // HlUsdcOracle oracle1;
@@ -19,7 +19,7 @@ contract DeployHyperliquidOracle is BaseScript {
 
         oracle = new HyperliquidOracle(asset, assetIndex, assetAmtScale, assetPriceScale);
         console2.log("HyperliquidOracle: ", address(oracle));
-        uint price = oracle.getValueInEth(asset, 1e18);
+        uint256 price = oracle.getValueInEth(asset, 1e18);
         console2.log("PURR/ETH: ", price);
 
         // oracle1 = new HlUsdcOracle();
